INTRODUCTION
termux it's a free app. It's available on the Internet. This they say hacking tool. We will use this to create a new and owned link. The link should then be shared on a social network that is relevant to the victim. Through it, we can keep track of where he is and where he lives.  This allows us to send a link wherever we know to someone who can easily find their details and where they are.

Location tracking has been of great importance since World War II, when military planners realized its usefulness for targeting, fleet management, positioning, and navigation [1]. Location tracking [2] is not one, single technology. Rather, it is the convergence of several technologies that can be merged to create systems that trackinventory, livestock or vehicle fleets. Similar systems can be created to deliver location-based services to wireless devices. Current technologies being used to create location-tracking and location-based systems include Geographic Information Systems (GIS), Global Positioning System (GPS), Radio Frequency Identification (RFID), and Wireless Local Area Network (WLAN). Location tracking or location-based service system will use one or a combination of these technologies. The system requires that a node or tag be placed on the object, animal or person being tracked. For example, the GPS receiver in a cell phone or an RFID tag on a DVD can be used to track those devices 
with a detection system such as GPS satellites or RFID receivers.