Create TERMUX APP
What is Termux?

Termux is a Linux Terminal Emulator application for Android. A terminal emulator is a program that allows the user to access the command line interface (CLI) in a graphical environment. If you’ve studied the basics of computer, you must know about the shell and command line interface. If not here is a short definition of shells for you to clear your concept.

What is Shell?

A shell is a command line interface, or you can say a program that takes commands from the user and instructs the operating system to perform tasks on the basis of these commands. There are many shells used in different operating systems like Bourne shell, C shell, Bourne Again shell. All have a different set of features and command behavior. Most of the Linux distributions have Bourne Again shell also known as bash as their default command line.
 

Why Termux?

Now we know that termux is a Linux Terminal Emulator. But why it is so famous and sometimes called a hacking machine. Well, that’s somehow true and sometimes a piece of false information also. Termux is not a standalone hacking application. As we discussed, it’s a program that takes instructions and runs scripts. In some Linux distributions like Kali, Parrot OS, we find a lot of tools for penetration testing. Most of them are CLI based and run through the terminal. We can also install various other tools and run different scripts made for Pentesting.
So on Android, obviously you get a complete GUI environment. Where these tools won’t work without any command line utility. That is where termux is used to install such tools on your Android Platform. Our Android is also based on a modified Linux kernel. That’s why we can use Bash scripting in it. That’s why termux can easily connect to the file system to perform the majority of the set of features. Which allows you to use the Linux command line to install and run various scripts.
So that was a wrap on introduction to Termux. Let’s move to installation and basic usage.

Installation:
There is no complicated process to install termux on your smartphone. Termux app is available on Google PlayStore. You can download the app and directly install it on your Android Phone. Also, there is an alternative way to install termux through F-Droid. It is actually the failsafe way to install termux with less built-in packages. The process is the same as you install any other application on your Android device. Just download, install and run the application.

Usage:

When you open the app, you’ll see a terminal interface. But Termux always gives you a basic graphical user interface. The interface is not that good but if you want to check the graphical environment in termux must use it. The GUI is based on the X11 Windows System.

Package management:

Termux provides APT and dpkg for package management and installation. Termux is similar in usage to most of the modern Linux distributions, especially Ubuntu and Debian based distros. The only major difference is that you won’t find the Filesystem Hierarchy Standard (FHS). The FHS is found in most of the Linux distributions. Which means you won’t find some directories on the same locations as they are in Linux distros.